Added comments to the examples to remove them later, added a small basis for the boat class and a potential boat manager
