--- RiverGame/Assets/Scripts/Boats/EnemyController.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyController : MonoBehaviour {
-	public int health;
-
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		if (health <= 0) {
-			Destroy(gameObject);
-		}
-	}
-}
