+++ RiverGame/Assets/Scripts/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 790525ef1fbd6b545b89ba1dfdb903d6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
