+++ RiverGame/Assets/Scripts/Enemy/EnemyController.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyController : MonoBehaviour
+{
+	public int health;
+	
+	public void Update()
+	{
+		if (health <= 0)
+		{
+			Destroy(gameObject);
+		}
+	}
+}
