+++ RiverGame/Assets/Scripts/Player/BoatManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffa9c86045508704c90701f2f34e3227
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
