+++ RiverGame/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.1.0f2
+m_EditorVersion: 2018.2.13f1
